BugFix: Button Commands Up / Down and Modules with ToolBars
This is a port of code that has already been accepted into the release_30
branch to get the Command Up and Down for buttons restored to Mudlet.

This provides the minimum to get the missing functionality back and
to fix a significant bug in the previous handling of toolbars that were
inserted into modules.

It DOES NOT fix the behaviour of push-down buttons not sending the right
thing on every press - a separate PR (also already inserted into the other
branch) will be prepared for that. A third PR that improves the layout of
the editor for Buttons/Menus/Toolbars and other items in the Editor has
prepared for the release_30 but has not been applied as it is not strictly
a bug fix.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>