Version: advance master to form start point for future work
It is my understanding that urgent bug fixes for the Release 3.0.0 need to
be based on a version with the same version major.minor.patch "Sematic
Version" string but, until actually "released" will carry a "-dev" build
suffix. This commit will introduce the change so that the resultant point
in the repository becomes the start point for such fixes and other work.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>