Resolve conflicts in src.pro [ci skip]
This sets the APP_VERSION in that file to 3.0.0 and the APP_BUILD to "-dev"
this matches the values in the CMake project file.

The .clang-format file is currently located in the ./src/ directory, that
has been used to select the appropriate file in the DISTFILES variable.

Note that there was missing HEADERS files mentioned in BOTH merge sources
for the irc sub-directory so I have picked through them to report the .h
that seem to be present.

Added ./CI/travis.set-build-info.sh and ./src/testdbg.h into QMake project
file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>