Resolve conflicts in TBuffer, TConsole, dlgMapper classes [ci skip]
Although unused (void)TBuffer::set_text_properties(int) was removed from
cpp file it was still declared in header file - so I whipped it out in
passing.

The TBuffer class also contains some currently commented out code to
provide HTML support for over-line and inverse which are both ANSI ESC
code-able SGR effects but which we do not currently provide.

The code changes for TConsole::toggleLogging(...) is slightly confusing at
first because some code was shifted to a separate
TConsole::slot_toggleLogging() wrapper used by the GUI but not by the Lua
subsystem that only calls the "wrapped" method.  This is so that on-console
logging starting and stopping messages could be avoided in the latter
usage, nevertheless the OTHER merge branch contained some
QStringLiteral(...) wrapped text strings that are better to use looking
forward to I18n.  Note that there are still some lines of text used to
build HTML source that also need such wrapping in the future.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>