Resolve remaining conflicts in: dlgProfilePreferences [ci skip]
A number of the conflicts in the .cpp file are repeated instances of code
to set a specific button to a user selected colour - the older conflicting
instances had some repeated boilerplate that A. Charles moved into a
separate setColor(Button, Color) function in the new instance - but which
did not take a local copy of the Host pointer and checked it for
non-nullness before it was used.  The resolution I have done here does
both.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>