Enhance: restore misfeature eliminated in Editor show/hide search results
In: https://github.com/Mudlet/Mudlet/commit/10649019416655df3e6499d83862b09e88f803ed
I attempted to give meaningful names to many elements that made up the main
Editor screen and to fix what looked like coding errors.  One of these was
the dodgy reuse of a pointer variable that was originally declared in the
trigger_editor.ui file and which pointed to the whole of the "Search" part
of the editor; it got reassigned to have the address of the "Results" part
only of what it originally pointed at.  The effect being that when its
hide() method was called only the "Results" part of the search area got
hidden.  I traced this behaviour back to the commit:
https://github.com/Mudlet/Mudlet/commit/fba7d68a77aa7b8eaf31891fba3bbf0ed69b2fe8
of 2009-11-12 17:0343 by Heiko which had the deeply meaningful commit
message of: "misc"!!!

My change returned the action of the button then called
"messageAreaCloseButton" which was (confusingly) connected to the SLOT
slot_show_search_area() and which I renamed then to "button_searchAreaClose"
to do what it seemed that it should do i.e. hide (it could not be used to
show the area as it was ON that area).

In the recent bug report:
https://bugs.launchpad.net/mudlet/+bug/1675988 "3.0 regression: search
field no longer stays visible" evidence was offered that some users wanted
the functionality of being able to hide just the results area of the search
function on the Editor.  This commit provides that by the addition of a
new SLOT that takes a boolean argument with the meaning of "show the search
results widget if true".  By connecting this to a new clickable tool-button
that I have positioned under the existing "close" button we can now offer
the wanted functionality.

To clarify the detail of this enhancement I added and revised some texts
that appear as tool- or status- tips.  Unfortunately in doing so it became
clear that there were some unused QActions / QMenu instances created in
the dlgTriggerEditor class constructor and that the at least one of them
had a tool-tip message that was erroneously being applied to a Editor
Toolbar button.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>