Revise: Editor Search feature always shown + fixup results toggle button
Following discussions it was decided to always show the search widget
so this commit removes the code that would show/hide it and the icon from
the Editor toolbar that is now not wanted.

During testing it was found that the triangular arrows that Qt provides as
an option for QToolbuttons is subject to StyleSheet styling and on current
MacOS platforms this causes the triangular icons were oversized.  As it was
not possible to locate the image files (if that is what they were) used to
make the triangles concerned I have taken the 48x48 pixel {GPL} Oxygen
icon files 1downarrow.png and 1rightarrow.png and converted them from green
to grey to use as replacement images and to limit the size of the button
to have a maximum square size of 0.75 of the height of the associated
QComboBox used for entry of search terms that the QToolButton is placed
alongside of.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>