Removed forceWebColorsTable() for the time being
Mudlet uses X11 colours as the base, not the W3C standard and it hasn't been an issue thus far - plus you can adjust the colours table to whatever is needed. Happy to put forceWebColorsTable() if there's demand for it but as far as I see there isn't much and having options people don't need is not sustainable long-term.