Revert: last commit and retry use of style-able built-in triangular icons
I found once I had undone some of the changes of the prior commit that I
might not have been using the right entity to use as the basis of the
size for the tool-button to show/hide the Editor search results - and that
I needed to force both the minimum and maximum sizes to use to ensure a
square resulting icon.

The triangular images I created for the prior commit have been removed
from the resource file. I have also remove the spacer that were used to
force the tool-button to the top-right of its area in the widget - it is
now centred which should give a sufficient margin around itself.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>