Update DB.lua
Changed check in db._migrate to "if type(cur) == "userdata" then" to match similar if statement later in the function. Added "or (not io.exists(getMudletHomeDir() .. "/Database_" .. db_name .. ".db"))" to check in db:create so that a new connection will be opened if the database file no longer exists.