Enhance: restore version check to Mudlet XMLimport code
I foresee the need to tweak the Xml format in the future, however though
there once was some checks for this in the code in the past it was
commented out.  This commit is a revised version of a pull request (#370)
originally aimed at the release_30 branch but which has changed so much it
was easier to restart from a more up to date point.

This version checked the "version" attribute of the MudletPackage element
of most types of XML files read (not Maps) and will allow through without
comment a string value that works out to a float less than 2.000f (three
decimal place resolution) but will cause XMLimport::importPackage(...)
to return a FALSE value - {the return is largely ignored currently) - to
not read the file further and to cause an "[ ALERT ]" message on the main
console should the version value be 2.000 or greater.

This is to cover the CURRENT code against problems that might arise in the
future if the XML format/structure gets changed too much which will
otherwise go *undetected* by the Mudlet application as it is currently.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>