Refactor: replace QIODevice with QFile pointer so can get file name
In XMLimport::importPackage(...) we would like to get access to the file
name details of the file being read but this is not available from the base
QIODevice type as it is a property of the QFile instances that are ALWAYS
used as the first argument to the method.  That being the case, to avoid
having to use a qobject_cast<T*>(...) or dynamic_cast<T*>(...) to convert
the original pointer to the desired type - with the need to handle
unanticipated situations where those might have failed, it makes sense to
explicitly use the QFile type so that we can use QFile::fileName() directly
as needed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>