Fixed GUI saved variables crashing with autologin
Mudlet was crashing when a profile was enabled for autologin and contained a saved variable. Issue was that code to initialise the 'base' TVar wasn't called as it was called in the non-autologin case (not sure why we have 2 functions for this?) and thus it was crashing when retrieving the name of the base variable.