Bugfix: make folder icons in edit have stable colouration
The colour used for "Keys" groups/folders is not consistent throughout the
Editor, they are set to Cyan (which is the same colour as "Menu" items in
the "Buttons" type), and Violet (which is the same colour as "Alias" items)
and Blue (which is the same colour as "Toolbars" in the "Buttons" type). To
resolve this I have created a pair of Pink icons (derived from the
existing icon files and thus also GPLed) and applied them uniformly as all
the relevant "Key" group item icons.

In passing I notice and re-did a couple of places that were missing
`{`...`}` around `if()` code blocks and a couple of places where nesting of
`else` and `if` branches were inconsistent in layout with other code
blocks.

Finally there were some places where if/else code was made with a condition
that had already been forced to one branch by a prior test in the same
region.  This involved a test for non-Nullness of a `pItem` pointer, so
as to minimise the noise in the git difference the redundant code has been
removed but the rest of the code layout left miss-indented.  This was
done in the knowledge that the entire C++/C code-base is to be "tidied"
up by a utility that will automatically reformat it all to a consistent
state in the very near future.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>