Enhance: add module/package install/uninstall event
This was one item inspired by the work of the coder(s) who forked Mudlet to
something that could handle Utf-8 in some manner. It was something that
struck me as potentially useful (they used it) but their implementation
did not quite match up to how we structure things.  The original only
reported the installation or un-installation of a module or package in
Host::installPackage(...) or Host::uninstallPackage(...)  but as those
methods are used in a number of manners it was remiss, I think, not to
report the second argument to those methods which may have a bearing on the
significance of them being run.  Also the original appended the package
or module name onto the end of the fixed event name meaning a separate
event handler would have to be registered to capture the loading or
unloading of each different package/module name concerned.  I have revised
the code so that a uniform "sysInstallPackageEvent" and
"sysUninstallPackageEvent" is generated for all modules/packages that are
changed, with the following arguments:

* "sysInstallPackageEvent"
* (string) package/module name
* (number) nature of event:
  * 0 - package
  * 1 - module (via UI or normal profile loading)
  * 2 - module (via a "sync" event from another profile)
  * 3 - module (via a script)
* (string) file name of package, archive file name for module

* "sysUninstallPackageEvent"
* (string) package/module name
* (number) nature of event:
  * 0 - package
  * 1 - module (via UI)
  * 2 - module (via a "sync" event from another profile)
  * 3 - module (via a script)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>