Change glColor4b to glColor4ub.
glColor4b uses signed bytes and represents the range -1 to 1.
glColor4ub uses unsigned bytes and represents the range 0 to 1.

I think the colors specified through glwidget.cpp are intended to
have 0 to 255 represent 0 to 1, as opposed to what actually
happened, which is that any value above 127 became negative
and represented a negative color value.