Revise: clean up some TEvent related stuff
Ensure all system generated events have event names that are NOT subjected
to translation.  However as I have recently learned that QLatin1String is
a better choice in most circumstances as opposed to QStringLiteral I have
changed to use them instead.

Rename some TEvent instances where our convention on naming suggested the
data structure reference was a pointed to a TEvent rather than the TEvent
itself (i.e. use of pE - at some stage in the past we were using pointers!)

Revamp TLuaInterpreter::setLabelClickCallback, setLabelReleaseCallback,
setLabelOnEnter and setLabelOnLeave to handle event arguments of boolean
and nil types - as we have expanded the range for TEvents everywhere else.
Also now returns a boolean true on success to find the given label or nil
and an error message on various run-time failures - including not finding
the label.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>