Revise: remove unwanted code source comments and use QLatin1String(...)
These events were inspired by "Zoilder" in a Mudlet derivative in 2015
called MudletRL for the Spanish MUD whose website is at
www.reinosdeleyenda.es

For "sysInstallPackage" the original version appended the packageName and
"Event" to "sysInstallPackage", but I have revised it to pass the
packageName as a separate argument. Due to the range of circumstances in
which this method can be called I have also included:
* the invoking "module" argument in case it is necessary to differentiate
  the way in which the installPackage code is called - I think this could
  be a factor in some use cases that Zoilder would not have experienced in
  their usage.
* the original package or module archive filename

The originators comment was effectively (ES):
«Se ejecuta el evento que indica que se ha instalado el paquete. Así el
propio paquete puede auto-inicializarse sin tener que reabrir la sesión».
which I think translates to (EN):
"The event that indicates that the package is installed runs. Thus the
package itself can auto-initialize without having to reopen the session."

For "sysUninstallPackageEvent" things are largely the same except that
there is no file name argument in the TEvent
The original comment was effectively (ES):
«Se ejecuta el evento que indica que se va a desinstalar el paquete. Así el
propio paquete puede eliminar lo que necesite sin tener que reabrir la
sesión».
Which translates, I think, to (EN):
"The event that indicates that the package is to be uninstalled is run. So
the package itself can eliminate what you need without having to reopen the
session."

I have recently learned that it is preferable to use QLatin1String rather
than QStringLiteral for QStrings that are not to be subjected to the
translation process!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>