BugFix: prevent clipping of Editor search term entry widget
There was a constraint set on a parent widget that prevented the QComboBox
that is used to enter search terms from re-sizing itself to accomodate
larger text than the form was originally designed for.

This commit removes that constraint and should prevent the clipping that
was reported to be occurring.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>