Revise: prevent translation of system event names
Ensure all system generated events have event names that are NOT subjected
to translation.  However as I have recently learned that QLatin1String is
a better choice in most circumstances as opposed to QStringLiteral I have
changed to use them instead.

As it is concerned with using QLatin1String wrappers around things
involving TEvent instances I have extended the re-coding to the qDebug()
code provided in TEvent.h to produce a formatted and useful dump of a
TEvent instance should it be passed to the Qt Debug system methods.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>