I18n: ensure Lua echo and feedTriggers commands pass Utf-8
Needed to allow some testing of other work related to I18n.  feedTriggers
did pass Utf-8 as it used (and still does) std::string to convey the data
without any conversion - however it needed a bit of an overhaul to conform
to the coding styles we now use.

The lua command echo is actually mapped to the TLuaInterpreter::Echo
function and had poor error messages - it also did not pass Utf-8.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>