I18n: clear out translatable Lua error messsages
Since it has been decided that the Mudlet Lua API should not have the
capability to report errors in using the API in a translated form the use
of the Qt tr(...) function to mark such QStrings is not wanted.  This
commit switches those usages to use the QStringLiteral(...) wrapper
instead. As the strings concerned are each unique the known downsides of
using this "non"-translatable QString wrapper should not be significant and
the code "template" for the two things are very similar {if there are no
second or second & third arguments in the tr(...) case} - the editing to
switch to QLatin1Strings in the same location would be much more intensive.

Where the (formerly) translated QString is then passed to the Lua system
there are some instances where QString::toUtf8() must be used instead of
QString::toLatin1() - these are generally the cases where a variable
included in the resulting QString can contain non-ASCII characters which
is any user defined/named item or {mainly on Windows platform} involving
file-system objects as that can also have non-ASCII characters.

There may be a few long text lines that I have manually inserted new-lines
into to try and prevent the resulting text from exceeding around 80
characters when shown on the main profile console.

There was a typo in one error message for TLuaInterpreter::downloadFile
as it described an issue for argument 1 when it should have said 2.  Also
in three functions a QString "windowName" is set to the null `QString` when
the `QString::clear()`is a little more descriptive of what is happening.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>