I18n: clear out translatable Lua error messsages (#941)
* I18n: clear out translatable Lua error messsages

Since it has been decided that the Mudlet Lua API should not have the
capability to report errors in using the API in a translated form the use
of the Qt tr(...) function to mark such QStrings is not wanted.  Indeed
the QString class type can be avoided entirely and we can return to
use raw literal strings with the lua_pushstring(...) and simplified C printf
format strings using %d for integer and %s for string replacement in
the related lua_pushfstring(...) function.

Where the values to be used as string replacements in lua_pushfstring
are themselves QStrings however there are some instances where
QString::toUtf8() must be used instead of QString::toLatin1() - these
are generally the cases where a variable included in the resulting
QString can contain non-ASCII characters which is any user
defined/named item or {mainly on Windows platform} involving
file-system objects as that can also have non-ASCII characters.

There may be a few long text lines that I have manually inserted new-lines
into to try and prevent the resulting text from exceeding around 80
characters when shown on the main profile console.

There was a typo in one error message for TLuaInterpreter::downloadFile
as it described an issue for argument 1 when it should have said 2.  Also
in three functions a QString "windowName" is set to the null `QString` when
the `QString::clear()`is a little more descriptive of what is happening.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>