Fixup: incorporate changes to TLuaInterpreter class into Pull-Request
Changes made in the series of commits making up this pull-request will
clash with other changes in TLuaInterpreter in a separate but now merged
pull-request initiated after this one was started.  By merging in the other
one and revising the changes in the third of the original commits in this
it will now be possible to merge the result back in to the development
branch without conflicts.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>