Fixed crash on centerview() with no map open. (#946)
* Fixed crash on centerview() with no map open.

Applied i18n guidelines and improved the message to be a lot more intuitive: it still concisely describes the problem yet also offers the obvious solution at once. I'm not sure if the `! pHost->mpMap->mpRoomDB` check is necessary anymore but it doesn't hurt to leave it in.

Fixes https://github.com/Mudlet/Mudlet/issues/945.