Enhance: add encoding/decoding to incoming and outgoing MUD data
Found during testing on the Spanish Realms of Legends MUD that we did not
support ISO-8859-1 (Latin-1) encoding but could if we made a small change
to the way we process the incoming text bytes in the
TBuffer::translateToPlainText(...) method.  Instead of appending each byte
as a char onto TBuffer::mMudline - which works for (7-bit) ASCII characters
we can convert the upper 128 code characters to the "correct" QChars by
going via a QChar::fromLatin1(...) intermediate step.

It was then found possible to extend this process for other simple encoding
tables by recording the differences between the ISO 8859-1 and various
other encodings.  Included in this commit is support for:
* ISO 8859-1
* ISO 8859-2
* ISO 8859-3
* ISO 8859-4
* ISO 8859-10
* ISO 8859-15
* ISO 8859-16
* WINDOWS-1250
* WINDOWS-1251
* WINDOWS-1252

The above (and UTF-8) can be selected from the profile preferences and will
be saved on a per profile basis once set and used when the profile is used
again. UTF-8 is manually decoded by examining the first byte to determine
how many more bytes are expected for it and then checking all of the
indicated bytes for validity.  Checks are included to guard against invalid
code points (beyond U+10FFFF) and "overlong" sequences.  During testing
against some data from a real MUD it was found that the UTF-8 BOM sequence
although not required or recommended for this encoding was being converted
to no QChars at all, technically the position in a stream where it occurs
could replace it with some form of joiner or it could actually be discarded
however a script or trigger might be interested in it so the equivalent
QChar is manually reinserted into the decoded stream - this is also useful
as the code structure wants to insert a TChar instance into the
corresponding data storage every time around the parsing loop for a single
byte (for all but UTF-8) encoding or Utf-8 byte sequence for a codepoint
(for UTF-8).

NOTE THAT THE DECODER WILL HANDLE NON-BMP BYTE SEQUENCES AND DECODES THEM
TO TWO QChars (a High surrogate followed by a Low one) AS QT DOES (It is
based on QStrings that are Utf-16). To allow code that follows this stage
to keep a one-to-one match between QChars in the text data and TChars that
record the rendering aspects (colour, font styling) a duplicate TChar is
added after the normal one for non-BMP codepoints.  I feel this will help
to port the existing code over to support such non-BMP codepoints...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>