Fixup: some revisions following reviews
Added getServerEncodingsList() which provides a list with the exact strings
to use - and which allows us to revised them if necessary.

Revise setServerEncoding() to not regard setting the same encoding that is
already being used from being a run-time error - now it will just return
the same true as if the setting had been made - but as no change is
actually taking place, no message will appear on the main console. This
message is now an "[ OK ] -" form one without an ending ellipses whereas it
was an "[ ALERT ] -" with one. Whether a change does occur or not a text
message is not provided, only on the run-time error case of specifying an
incorrect encoding will a nil + error-message now be produced.

Comment at start of TBuffer::translateToPlainText(...) redone.

Renamed data input argument to above function to ensure that the local
version which can be slightly longer if there are bytes "left-over" from
the previous packet that have been prepended to it.  Doing this allowed me
to find and fix places where that was not being used.

Revised some debugging messages to be clearer.

Revised the error handling for where the decoded UTF-8 data yields more
than two QChars from when converted by a QString constructor - it should
not happen but being paranoid is not unreasonable with initial coding
attempts...! Also arranged that more error cases report the actual bytes
that seemed to be in error.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>