Revise: use hard coded tables for decoding MUD Server data
Now using data recovered from Qt's own code repository, and adds additional
encodings - which could be subsequently added/removed as we see fit.

Code also restructured to work with a singleton(?) table of the above data,
in the form of a QMap<QString, QVector<QChar>> where the QVector is the
128 Unicode code-point values for the characters in the range 128-255 for
each of the Extended ASCII tables.  This excludes ISO 8859-1 as that is
implicitly the one-to-one mapping 128=128, 129=129,... 255=255 and
obviously excludes ASCII which does not go that far and UTF-8 which is
handled differently.  It also allows a significant performance improvement
as it eliminates a pair of switch() constructs and a call to a encoding
specific method for each byte of data for the other encoding and replaces
them with a single look-up in a single QVector; that being initialised at
the start of processing the packet of data.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>