Rebased revise ensure sys event names not translated (#934)
* Revise: prevent translation of system event names

Ensure all system generated events have event names that are NOT subjected
to translation.  However as I have recently learned that QLatin1String is
a better choice in most circumstances as opposed to QStringLiteral I have
changed to use them instead.

As it is concerned with using QLatin1String wrappers around things
involving TEvent instances I have extended the re-coding to the qDebug()
code provided in TEvent.h to produce a formatted and useful dump of a
TEvent instance should it be passed to the Qt Debug system methods.

* Revise: rationalise variable names as TEvent instances

Rename some TEvent instances where our convention on naming suggested the
data structure reference was a pointed to a TEvent rather than the TEvent
itself (i.e. use of pE - at some stage in the past we were using pointers!)

In the light of revisions to our code style choices - I also adopt those in
the areas I have modified.

* Refactor: add nil/bool value type support to Label callback functions

We recently added nil and boolean type argument support to the TEvent
infrastructure however it was missed from being added to labels.  This will
do that and also bring the coding styles recently decided upon to those Lua
functions.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>