Revise: correct small UTF-8 decoder bug and revise encoding setting tooltip
This fixes a coding bug that would lose 1 to 3 bytes of server data should
a UTF-8 byte stream from the MUD server be fragmented between two incoming
network data packets.

It also adds a warning in the tool-tip of the encoding setting control to
leave the encoding setting on the profile preferences to be ASCII or
ISO 8859-1 as others will not yet work with the remainder of the
application's code base.  Some phrasing in the somewhat lengthy text of
that tool-tip (suggestions are being considered as to "precis"ing it to a
smaller version!) have also been made.

The name of the setting has also been change from "Text encoding" to
"Server Data Encoding" to (I feel) better indicate it's purpose.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>