BugFix: allows most ASCII control characters to be used/saved in Lua code
We use XML to save Mudlet game data including the scripts that contain
Lua code for all of the Mudlet item {Alias|Button|Key|Timer|Script|Trigger}
but because the former (at version 1.0 which is all that the Qt library
code handles) prohibits all but Horizontal Tab, Carriage Return and Line
Feed out of the range of ASCII control codes it means that a user trying to
embed a raw string containing, say, the ESC code (0x1b) will either lose
their code for just the item containing that code OR everything that
follows it in the file that it is being loaded from.  As that could be the
game save data this may will cause data loss.

This commit allows all but the ASCII NULL (`\0`) character code to be
stored as it replaces the remaining control codes with a pair of other
Unicode code-points that can be safely stored, that is (U+FFFC) the Unicode
Object Replacement Character code-point followed by one of the code-points
from the Control Picture Symbol range. The former is not a visible
character in normal circumstances but the latter (if present in a font) is
typically a two or three letters in a diagonal line in a single grapheme
that has the same two or three letters used to abbreviate an ASCII Control
code in, say, visible table representations.  The chances of the Object
Replacement Character occurring in ANY document is virtually zero so the
use for this purpose should be "safe" in this context and using the Control
Picture Character does have the nice feature in that viewing a saved file
with them in does suggest which character is being used - and the chances
of them being used in, even a Mudlet script, otherwise is vanishingly small
in my opinion.

Note that NO attempt is made to handle the ASCII NUL character as that is
also, symbolically the end of a C/C++ string normally and it would be
confusing to the Lua interpreter even though the Qt QString text handling
system could tolerate having such characters in positions other than at the
end of a QString.

During the development of this system to "Escape" these codes I did
experiment with the use of "XML Entities" to represent them - such as
"&ESC;" to stand in for the ASCII 0x1B character.  However the Qt process
to insert such elements into the required a costly splitting the script
text up into fragments without any of the codes to be escaped so that each
code could then be injected into the QXmlStreamWriter stream interleaved
with the text fragments that surrounds them and whilst this proved to be
possible I then found that the process that then took the entities and
substituted in the original ASCII control codes when the script data was
read with the QXmlStreamReader failed to work because the
QXmlStreamEntityResolver applied the same character code restrictions onto
the replacement text - which is exactly not what was wanted!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>